// src/Type.ts
export enum Type {
  Bool = 1,
  Null = 2,
  Undefined = 3,
  U8 = 4,
  U16 = 5,
  U32 = 6,
  U64 = 7,
  I8 = 8,
  I16 = 9,
  I32 = 10,
  I64 = 11,
  Int = 12,
  F16 = 13,
  F32 = 14,
  F64 = 15,
  Simple = 16,
  Bytes = 17,
  BytesIndef = 18,
  String = 19,
  StringIndef = 20,
  Array = 21,
  ArrayIndef = 22,
  Map = 23,
  MapIndef = 24,
  Tag = 25,
  Break = 26,
}